+++ providers/admin_object.rb
@@ -16,7 +16,9 @@
 
 include Chef::Asadmin
 
-notifying_action :create do
+use_inline_resources
+
+action :create do
 
   command = []
   command << "create-admin-object"
@@ -37,7 +39,7 @@ notifying_action :create do
   end
 end
 
-notifying_action :delete do
+action :delete do
   command = []
   command << "delete-admin-object"
   command << asadmin_target_flag