+++ providers/auth_realm.rb
@@ -16,7 +16,9 @@
 
 include Chef::Asadmin
 
-notifying_action :create do
+use_inline_resources
+
+action :create do
   command = []
   command << "create-auth-realm"
   command << asadmin_target_flag
@@ -35,7 +37,7 @@ notifying_action :create do
   end
 end
 
-notifying_action :delete do
+action :delete do
   command = []
   command << "delete-auth-realm"
   command << asadmin_target_flag