+++ providers/custom_resource.rb
@@ -16,7 +16,9 @@
 
 include Chef::Asadmin
 
-notifying_action :create do
+use_inline_resources
+
+action :create do
   command = []
   command << "create-custom-resource"
   command << "--restype" << new_resource.restype
@@ -56,7 +58,7 @@ notifying_action :create do
   end
 end
 
-notifying_action :delete do
+action :delete do
   command = []
   command << "delete-custom-resource"
   command << asadmin_target_flag