+++ providers/deployable.rb
@@ -17,7 +17,9 @@ require 'digest/sha1'
 
 include Chef::Asadmin
 
-notifying_action :deploy do
+use_inline_resources
+
+action :deploy do
   version_value = new_resource.version ? new_resource.version.to_s : Digest::SHA1.hexdigest(new_resource.url)
   base_cache_name = "#{Chef::Config[:file_cache_path]}/#{new_resource.domain_name}_#{new_resource.component_name}_#{version_value}"
   versioned_component_name = Asadmin.versioned_component_name(new_resource.component_name, new_resource.version, new_resource.url, new_resource.descriptors)
@@ -119,7 +121,7 @@ test -f #{deployment_plan}
   end
 end
 
-notifying_action :undeploy do
+action :undeploy do
   command = []
   command << "undeploy"
   command << "--cascade=true"
@@ -134,7 +136,7 @@ notifying_action :undeploy do
   end
 end
 
-notifying_action :disable do
+action :disable do
   command = []
   command << "disable"
   command << asadmin_target_flag
@@ -148,7 +150,7 @@ notifying_action :disable do
   end
 end
 
-notifying_action :enable do
+action :enable do
   command = []
   command << "enable"
   command << asadmin_target_flag