+++ providers/mq_destination.rb
@@ -16,7 +16,9 @@
 
 include Chef::Imqcmd
 
-notifying_action :create do
+use_inline_resources
+
+action :create do
   Chef::Log.info "Creating MQ Destination #{new_resource.destination_name}"
 
   bash "imqcmd_create_#{new_resource.queue ? "queue" : "topic"} #{new_resource.destination_name}" do
@@ -45,7 +47,7 @@ notifying_action :create do
 end
 
 
-notifying_action :destroy do
+action :destroy do
   bash "imqcmd_create_#{new_resource.queue ? "queue" : "topic"} #{new_resource.destination_name}" do
     only_if "#{imqcmd_command("query dst -t #{new_resource.queue ? 'q' : 't'} -n #{new_resource.destination_name}")} >/dev/null"
     user node['glassfish']['user']