+++ providers/property.rb
@@ -16,7 +16,9 @@
 
 include Chef::Asadmin
 
-notifying_action :set do
+use_inline_resources
+
+action :set do
   bash "asadmin_set #{new_resource.key}=#{new_resource.value}" do
     not_if "#{asadmin_command("get #{new_resource.key}")} | grep -x -- '#{new_resource.key}=#{new_resource.value}'"
     user node['glassfish']['user']