+++ providers/resource_adapter.rb
@@ -16,7 +16,9 @@
 
 include Chef::Asadmin
 
-notifying_action :create do
+use_inline_resources
+
+action :create do
 
   command = []
   command << "create-resource-adapter-config"
@@ -35,7 +37,7 @@ notifying_action :create do
   end
 end
 
-notifying_action :delete do
+action :delete do
   command = []
   command << "delete-resource-adapter-config"
   command << new_resource.resource_adapter_name