+++ providers/secure_admin.rb
@@ -16,7 +16,9 @@
 
 include Chef::Asadmin
 
-notifying_action :enable do
+use_inline_resources
+
+action :enable do
   service "glassfish-#{new_resource.domain_name}" do
     provider Chef::Provider::Service::Upstart
     supports :restart => true, :status => true
@@ -32,7 +34,7 @@ notifying_action :enable do
   end
 end
 
-notifying_action :disable do
+action :disable do
   service "glassfish-#{new_resource.domain_name}" do
     provider Chef::Provider::Service::Upstart
     supports :restart => true, :status => true